Use a config file for test settings
Instead of configuring entirely from the command-line, use a config
file.  Settings related to the contents of the test (like which
operations are enabled) go in the file.  Settings related to the
execution of the test (like how many steps to run) are given on the
command line.

This change adds the config file and moves some settings over, but not
all.